new_project/requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
mysqlclient==2.2.7
PyMySQL==1.1.2
sqlparse==0.5.3
tzdata==2025.2